requirements.txt
Django==1.6.2
PIL==1.1.7
South==0.8.4
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-autocomplete-light==2.0.0a15
django-debug-toolbar==1.0.1
django-tagging==0.3.2
django-toolbelt==0.0.1
gunicorn==18.0
imagestore==2.7.6
psycopg2==2.5.2
pystache==0.5.3
six==1.5.2
sorl-thumbnail==11.12
sqlparse==0.1.11
static==1.0.2
wsgiref==0.1.2
